requirements.txt
pyTelegramBotAPI==4.22.0
tavily-python==0.5.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.24.3